Cold Email Generator/requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
chromadb
pandas
python-dotenv
beautifulsoup4
lxml
flask
flask-cors